cleanup code, add authenticate controller
