--- server/config/passport.js
@@ -1,8 +1,0 @@
-var passport = require('passport');
-var OAuth2Strategy = require('passport-oauth').OAuth2Strategy;
-var User = require('../models/user.js');
-
-var liStrategy = new OAuth2Strategy({
-
-
-});
