+++ todo.todo
@@ -10,9 +10,9 @@ OBJECTIVES:
    ✔ Cancel - redirect #/ @done (15-05-08 09:09)
  ☐ Open public route to get cards
    ✔ Create Card collection @done (15-05-08 10:58)
-   ☐ add card reference to user object
-   ☐ populate cards on request
+   ✔ add card reference to user object @done (15-05-08 12:00)
+   ✔ populate cards on request @done (15-05-08 12:00)
 
+ ☐ Delete cards
  ☐ Add profile picture to public card
  ☐ Float users cards
- ☐ Delete cards
